blog/requirements/local.txt
asgiref==3.2.7
Django==3.0.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-model-utils==4.1.1
Pillow==8.1.0
psycopg2==2.8.6
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1